# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List approvals (optionally filtered by charter or approver)
      parameters:
        - name: charterId
          in: query
          required: false
          schema:
            type: integer
        - name: approverId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
  /approvals/{id}:
    get:
      operationId: getApproval
      tags: [approvals]
      summary: Get an approval by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/decide:
    post:
      operationId: decideApproval
      tags: [approvals]
      summary: Approve or reject an approval step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionBody"
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
  /dashboard/pending-approvals:
    get:
      operationId: getPendingApprovals
      tags: [dashboard]
      summary: Get all pending approvals across all charters
      parameters:
        - name: approverId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Pending approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
...
  /projects/{id}/budget-lines:
    get:
      operationId: listBudgetLines
      tags: [budget]
      summary: List budget lines for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of budget lines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetLine"
    post:
      operationId: createBudgetLine
      tags: [budget]
      summary: Create a budget line for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetLineBody"
      responses:
        "201":
          description: Created budget line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetLine"
  /budget-lines/{id}:
    patch:
      operationId: updateBudgetLine
      tags: [budget]
      summary: Update a budget line
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetLineBody"
      responses:
        "200":
          description: Updated budget line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetLine"
    delete:
      operationId: deleteBudgetLine
      tags: [budget]
      summary: Delete a budget line
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Approval:
      type: object
      properties:
        id:
          type: integer
        charterId:
          type: integer
        approverId:
          type: ["integer", "null"]
        approverRole:
          type: string
          description: "Role: hod, executive_director, cfo, scm, chairman, finance, pmo"
        stage:
          type: string
          description: "Stage: parallel_review, scm_review, chairman_review, finance_review, pmo_review"
        status:
          type: string
          description: "Status: pending, approved, rejected"
        comments:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        charterTitle:
          type: ["string", "null"]
        approverName:
          type: ["string", "null"]
      required:
        - id
        - charterId
        - approverRole
        - stage
        - status
        - createdAt
    ApprovalDecisionBody:
      type: object
      properties:
        decision:
          type: string
          description: "Decision: approved or rejected"
        comments:
          type: string
      required:
        - decision
...
    BudgetLine:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        category:
          type: string
          description: "Category: CapEx, OpEx"
        description:
          type: string
        baselineAmount:
          type: number
        forecastAmount:
          type: number
        actualAmount:
          type: number
        varianceAmount:
          type: number
        variancePct:
          type: number
        period:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - projectId
        - category
        - baselineAmount
        - forecastAmount
        - actualAmount
        - varianceAmount
        - variancePct
        - createdAt
    CreateBudgetLineBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        baselineAmount:
          type: number
        forecastAmount:
          type: number
        actualAmount:
          type: number
        period:
          type: string
    UpdateBudgetLineBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        baselineAmount:
          type: number
        forecastAmount:
          type: number
        actualAmount:
          type: number
        period:
          type: string
...