# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Approval {
  id: number;
  charterId: number;
  /** @nullable */
  approverId?: number | null;
  /** Role: hod, executive_director, cfo, scm, chairman, finance, pmo */
  approverRole: string;
  /** Stage: parallel_review, scm_review, chairman_review, finance_review, pmo_review */
  stage: string;
  /** Status: pending, approved, rejected */
  status: string;
  /** @nullable */
  comments?: string | null;
  /** @nullable */
  decidedAt?: string | null;
  createdAt: string;
  /** @nullable */
  charterTitle?: string | null;
  /** @nullable */
  approverName?: string | null;
}